% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_convert_rfus.R
\name{ce_convert_rfus}
\alias{ce_convert_rfus}
\title{Converts RFUs to concentration in μg/L}
\usage{
ce_convert_rfus(
  rfu_in,
  module = c("ext_chla", "invivo_chla", "phyco"),
  year = years,
  fluorometer = c("ours", "theirs"),
  output = NULL
)
}
\arguments{
\item{rfu_in}{A .csv file with sample metadata and RFU.  See ADDEXAMPLEFILE
for an example of the expected file format.}

\item{module}{One of two options: "ext_chla", "invivo_chla", or "phyco", 
default is "ext_chla".}

\item{year}{Year of the standard curve.  If more than one curve caluclated in
a year add "a", "b", ...  Acceptable values are:  "2021"}

\item{fluorometer}{One of two options: "ours" or "theirs", default is 
"ours". These represent the two Turner Triology's at
ACESD.  The "ours" one was purchased by the compeco lab, 
the "theirs", has been in use for sometime and currently 
resides in M07. The default is "ours".}

\item{output}{An options output path and csv file name for converted values.}
}
\value{
returns a tibble with proper metadata, input RFUs and converted
         concentrations
}
\description{
This functions needs input data from a fluorometer of RFU's and then
uses a standard curve to make the conversion.  The standard curves are stored
internally in this package and may be updated.  The module, fluorometer, and 
date arguments will specify which curve data file to use.
}
\note{
While it is possible to mix and match modules and fluorometers, don't 
       do this.  Keep the "ours" modules with the "ours" fluorometer 
       and vice-versa.
}
\examples{
examp_data <- system.file("extdata/chla_2021_7_29.csv", package = "compeco")
ce_convert_rfus(rfu_in = examp_data)
}
