% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_convert_rfus.R
\name{ce_convert_to_conc}
\alias{ce_convert_to_conc}
\title{Convert RFUS to concentration}
\usage{
ce_convert_to_conc(
  module = c("ext_chla", "invivo_chla", "phyco"),
  rfus,
  std_curve,
  blank_correction = TRUE
)
}
\arguments{
\item{module}{which module/variable}

\item{rfus}{rfu object from \code{ce_convert_rfus}}

\item{std_curve}{std_curve object from \code{ce_convert_rfus}}
}
\description{
This function creates takes an compeco input rfu object and standard curve 
object and converts the rfus to concentration
}
\keyword{internal}
