% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_convert_rfus.R
\name{ce_create_std_curve}
\alias{ce_create_std_curve}
\title{Create a standard curve}
\usage{
ce_create_std_curve(...)
}
\arguments{
\item{...}{Arguments for module, fluorometer and date of input standard curve
csv files.  Passed from \code{\link{ce_convert_rfus}}}
}
\description{
This function creates a standard curve from input fluorometer and 
spectrophotometer .csv files.  Curve is used to convert RFUs to µg/L.
}
\keyword{internal}
