% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_convert_rfus.R
\name{ce_export_std_curve}
\alias{ce_export_std_curve}
\title{Export Standard Curve}
\usage{
ce_export_std_curve(module, fluorometer, year)
}
\arguments{
\item{module}{One of two options: "ext_chla", "invivo_chla", or "phyco", 
default is "ext_chla".}

\item{fluorometer}{One of two options: "g04" or "m07", default is 
"g04". These represent the two Turner Triology's at
ACESD.  The "g04" one was purchased by the compeco lab, 
the "m07", has been in use for sometime and currently 
resides in M07. The default is "g04".}

\item{year}{Year of the standard curve.  If more than one curve caluclated in
a year add "a", "b", ...  Acceptable values are:  "2021" for
all but ext_chla and g04.  That has "2021" and "2022".}
}
\description{
This function exports the linear model for the different standard curves.
}
\examples{
ce_export_std_curve("ext_chla", "g04", "2022")
}
