% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_process_buoy_data.R
\name{ce_process_buoy_data}
\alias{ce_process_buoy_data}
\title{Deal with buoy data}
\usage{
ce_process_buoy_data(
  remote_data = c("sftp://newftp.epa.gov/buoys/"),
  user = Sys.getenv("NEWFTPU"),
  pass = Sys.getenv("NEWFTPP"),
  output = here::here("data/merged_buoy_data.feather")
)
}
\description{
This function fetches buoy data from a remote url (currently EPA's newftp), 
downloads it, cleans it, and merges it with existing data.
}
\details{
@param remote_data URL for the remote data location, currently using 
                    newftp.epa.gov
 @param user User name for accessing the remote-data URL, defaults to local 
             environment variable, NEWFTPU
 @param pass Password for accessing the remote-data URL, defaults to local 
             environmnetal variable, NEWFTPP.  Ok to use interactively with 
             user supplied password, but be careful with passwords stored in
             scripts using this function.  Instead use enviornment variables.
 @param output Name of output .feather file.  Using arrow::write_feather as 
               file sizes are great and read/write is super fast!
 @return Tibble of merged data of all buoy data.
 @export
}
