% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_read_exo_data.R
\name{ce_read_exo_data}
\alias{ce_read_exo_data}
\title{Reads and cleans an EXO2 data file}
\usage{
ce_read_exo_data(exo_file, source = c("handheld", "korexo"))
}
\arguments{
\item{exo_file}{Path and file name for an exo file to read}

\item{source}{How were the data retrieved from the sensor.  Current options 
include "handheld" or "korexo"}
}
\description{
Data files pulled from a YSI EXO sonde come from either a YSI handheld or 
via KorEXO.  This function will read those files and convert them into a 
data frame.
}
\examples{
exo_file_path <- system.file("extdata/210915-112052-_.csv", package = "compeco")
ce_read_exo_data(exo_file_path)
}
