% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_process_buoy_data.R
\name{clean_buoy_data}
\alias{clean_buoy_data}
\title{clean buoy data}
\usage{
clean_buoy_data(csv_file)
}
\description{
This is Cape Code specific now.  Will work until we add other buoys.  Need 
to think about how to make this more universal.  Might be on the buoy to 
web side of things.
}
\keyword{internal}
