% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_get_nla.R
\name{clean_nla_2007}
\alias{clean_nla_2007}
\title{Clean 2007 NLA Water Chemistry Data}
\usage{
clean_nla_2007(wc_chla_df, tox_df)
}
\arguments{
\item{wc_chla_df}{A data frame of 2007 NLA water chemistry and chlorphyll 
data, from \code{get_nla_2007_waterchem()}.}

\item{tox_df}{A data frame of 2007 NLA cyanotoxin data from 
\code{get_nla_2007_toxin()}.}
}
\value{
A tibble of cleaned 2007 NLA site data
}
\description{
Clean 2007 NLA Water Chemistry Data
}
