% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_get_nla.R
\name{clean_nla_2012}
\alias{clean_nla_2012}
\title{Clean 2012 NLA Water Chemistry Data}
\usage{
clean_nla_2012(wc_df, site_df, chla_df, tox_df)
}
\arguments{
\item{wc_df}{A data frame of NLA water chemistry data, from 
\code{get_nla_2012_waterchem()}.}

\item{site_df}{A data frame of 2012 NLA sites, from 
\code{clean_nla_2012_site()}.}

\item{chla_df}{A data frame of 2012 NLA chlorophyll data, from 
\code{get_nla_2012_chla()}.}

\item{tox_df}{A data frame of 2012 NLA cyanotoxin data, from 
\code{get_nla_2012_toxin()}.}
}
\value{
A tibble of cleaned 2012 NLA site data
}
\description{
Clean 2012 NLA Water Chemistry Data
}
