% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_get_nla.R
\name{clean_nla_2012_site}
\alias{clean_nla_2012_site}
\title{Clean 2012 NLA Site Data}
\usage{
clean_nla_2012_site(site_df)
}
\arguments{
\item{site_df}{A data frame of NLA site data, from 
\code{get_nla_2012_site()}.}
}
\value{
A tibble of cleaned 2012 NLA site data
}
\description{
Clean 2012 NLA Site Data
}
