% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_get_nla.R
\name{get_nla_2007_all}
\alias{get_nla_2007_all}
\title{Function to get 2007 All Data Zip}
\usage{
get_nla_2007_all(data_dir = "data", zip_dir = "nla2007")
}
\arguments{
\item{data_dir}{Location to download and save data to. Defaults to a "data" 
folder in the current working directory.  Will create if it
does not current exist.}

\item{zip_dir}{location, within data_dir to extract files.  Defaults to 
"nla2007"}
}
\value{
Character path for unzipped data
}
\description{
Downloads and extracts all 2007 NLA data
}
