% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_get_nla.R
\name{get_nla_2007_site}
\alias{get_nla_2007_site}
\title{Function to get 2007 NLA Site data}
\usage{
get_nla_2007_site(
  nla_site_file = "data/nla2007/NLA2007_SampledLakeInformation_20091113.csv"
)
}
\arguments{
\item{nla_site_file}{path to where nla data exist.  Defaults to locally 
downloaded files in "data/nla2007"}
}
\value{
A tibble
}
\description{
Grabs data from a path, switches all columns to lower case.
}
