% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_get_nla.R
\name{get_nla_2007_toxin}
\alias{get_nla_2007_toxin}
\title{Function to get 2007 NLA toxin data}
\usage{
get_nla_2007_toxin(
  nla_rec_cond_file = "data/nla2007/NLA2007_Recreational_ConditionEstimates_20091123.csv"
)
}
\arguments{
\item{nla_rec_cond_file}{path to where nla data exist.  Defaults to locally 
downloaded files in "data/nla2007"}
}
\value{
A tibble
}
\description{
Grabs data from a path, switches all columns to lower case.  The recreational 
condition data includes toxins and various other measures.
}
