% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_get_nla.R
\name{get_nla_2007_wc_chla}
\alias{get_nla_2007_wc_chla}
\title{Function to get 2007 NLA water quality data}
\usage{
get_nla_2007_wc_chla(
  nla_wq_file = "data/nla2007/NLA2007_WaterQuality_20091123.csv"
)
}
\arguments{
\item{nla_wq_file}{path to where nla data exist.  Defaults to locally 
downloaded files in "data/nla2007"}
}
\value{
A tibble
}
\description{
Grabs data from a path, switches all columns to lower case.
}
