% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_get_nla.R
\name{get_nla_2012_site}
\alias{get_nla_2012_site}
\title{Function to get 2012 NLA Site data}
\usage{
get_nla_2012_site()
}
\value{
A tibble
}
\description{
Grabs data from EPA site, switches all columns to lower case, removes 
unsampled sites.
}
