% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_get_nla.R
\name{get_nla_2012_wc}
\alias{get_nla_2012_wc}
\title{Function to get 2012 NLA water chemistry data}
\usage{
get_nla_2012_wc()
}
\value{
A tibble
}
\description{
Grabs data from EPA site, switches all columns to lower case.
}
