% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_get_nla.R
\name{merge_nla}
\alias{merge_nla}
\title{Merge multiple NLAs}
\usage{
merge_nla(...)
}
\arguments{
\item{...}{NLA data frames to merge.  These are output from 
\code{clea_nla_xxx()}.}
}
\value{
A merged tibble of cleaned NLA data
}
\description{
This function takes cleaned up NLA datasets and merges them into a single
master NLA data frame
}
