% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_convert_rfus.R
\name{ce_convert_rfus}
\alias{ce_convert_rfus}
\title{Converts RFUs to concentration in μg/L}
\usage{
ce_convert_rfus(
  rfu_in,
  module = c("ext_chla", "invivo_chla", "phyco"),
  fluorometer = c("compeco", "acesd"),
  date = c("2021-11-16"),
  output = NULL
)
}
\arguments{
\item{rfu_in}{A .csv file with sample metadata and RFU.  See ADDEXAMPLEFILE
for an example of the expected file format.}

\item{module}{One of two options: "ext_chla", "invivo_chla", or "phyco", 
default is "ext_chla".}

\item{fluorometer}{One of two options: "compeco" or "acesd", default is 
"compeco". These represent the two Turner Triology's at
ACESD.  The "compeco" one was purchased by our lab, the
"acesd", has been in use for sometime and currently 
resides in M07. The default is "compeco".}

\item{date}{Date of the standard curve.  Acceptable values are:  "NEED THESE"}

\item{output}{An options output path and csv file name for converted values.}
}
\value{
returns a tibble with proper metadata, input RFUs and converted
         concentrations
}
\description{
This functions needs input data from a fluorometer of RFU's and then
uses a standard curve to make the conversion.  The standard curves are stored
internally in this package and may be updated.  The module, fluorometer, and 
date arguments will specify which curve data file to use.
}
\note{
While it is possible to mix and match modules and fluormeters, don't 
       do this.  Keep the "compeco" modules with the "compeco" fluorometer 
       and vice-versa.
}
\examples{
ce_convert_rfus()
}
