% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_convert_rfus.R
\name{ce_convert_rfus}
\alias{ce_convert_rfus}
\title{Converts RFUs to concentration in μg/L}
\usage{
ce_convert_rfus(
  rfus,
  module = c("ext_chla", "invivo_chla", "phyco"),
  year = c("2021", "2022"),
  fluorometer = c("g04", "m07"),
  output = NULL,
  std_check = TRUE,
  conversion_slope = NULL
)
}
\arguments{
\item{rfus}{A data frame with sample metadata and RFU.  See ADDEXAMPLEFILE
for an example of the expected file format.}

\item{module}{One of two options: "ext_chla", "invivo_chla", or "phyco", 
default is "ext_chla".}

\item{year}{Year of the standard curve.  If more than one curve caluclated in
a year add "a", "b", ...  Acceptable values are:  "2021"}

\item{fluorometer}{One of two options: "g04" or "m07", default is 
"g04". These represent the two Turner Triology's at
ACESD.  The "g04" one was purchased by the compeco lab, 
the "m07", has been in use for sometime and currently 
resides in M07. The default is "g04".}

\item{output}{An optional output path and csv file name for converted values.}

\item{std_check}{Logical to determine if a solid standard check should be 
done.  Defualts to TRUE.}

\item{conversion_slope}{This argument is used to overide the standard curves 
that are generated internally.  If you are using a 
different fluorometer than specified above and 
already know the slope, specify that here as a 
numeric.}
}
\value{
returns a tibble with proper metadata, input RFUs and converted
         concentrations
}
\description{
This functions needs input data from a fluorometer of RFU's and then
uses a standard curve to make the conversion.  The standard curves are are
calculated and stored internally in this package and may be updated.  
Fluorometer and spec files needed for the standard curves are detailed in 
ADDSOPHERE and will work on both blanked and unblanked spec measurements.  
Note that the ADDSOPHERE SOP sepcifies blanked mesaurements.  The unblanked 
measurements are include for older standard curve methods but should not be 
used going forward. The module, fluorometer, and date arguments will specify 
which curve data file to use.
}
\note{
While it is possible to mix and match modules and fluorometers, don't 
       do this.  Keep the "g04" modules with the "g04" fluorometer 
       and vice-versa.
}
\examples{
examp_chla_data <- readr::read_csv(system.file("extdata/chla_2021_7_14.csv", package = "compeco"), na = c("", "NA", "na"))
examp_phyco_data <- readr::read_csv(system.file("extdata/phyco_2021_06_28.csv", package = "compeco"), na = c("", "NA", "na"))
surge_chla <- readr::read_csv(system.file("extdata/surge_file.csv", package = "compeco"), na = c("", "NA", "na"))

ce_convert_rfus(rfus = examp_chla_data, module = "ext_chla", year = "2021", fluorometer = "m07")
ce_convert_rfus(rfus = examp_phyco_data, module = "phyco", fluorometer = "g04")
ce_convert_rfus(rfus = surge_chla, year = "2022", module = "ext_chla", 
                fluorometer = "g04", std_check = FALSE)
}
